A simple "Hello World" REST Demo
This is a simple "Hello world" type application to showcase the use of
the XMS Rest Interface.
http://www.dialogic.com/webhelp/XMS/2.3/XMS_RESTfulAPIUser.pdf
The source files are commented to show the usage of the Appache HTTP
libraries to interface with the XMS.

This application will
- Allocate an object (XMSEventListener) to establish the connection via
a LongPoll.  This will wait for events in the HttpStream and then notify
all observers on the reception of a new event

- Create an object that can be used to generate and send the Http
request (XMSRestSender)

- Use the XMSRestSender to create a new call via POST
- Wait for the Stream to establish and notification from the
EventListener
- Play out a file to the call by using the XMSRestSender and a PUT
- Wait for the END_PLAY notification from the Event Listener
- Hangup/Terminate the call via a DELETE